Relocate esphome entity code into its own module
