+++ homeassistant/components/esphome/binary_sensor.py
@@ -13,8 +13,12 @@ from homeassistant.core import HomeAssistant, callback
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.util.enum import try_parse_enum
 
-from . import EsphomeAssistEntity, EsphomeEntity, platform_async_setup_entry
 from .domain_data import DomainData
+from .entity import (
+    EsphomeAssistEntity,
+    EsphomeEntity,
+    platform_async_setup_entry,
+)
 
 
 async def async_setup_entry(
